test: run concurrent share lock test in unit CI for coverage
